4_detector_huggingface/requirements.txt
ultralytics
opencv-python-headless
streamlit
Pillow